group部分修改

+++ DevOpsTeamwork/src/main/java/teamworkbranch/module/group/service/impl/GroupServiceImpl.java
@@ -3,6 +3,8 @@ package teamworkbranch.module.group.service.impl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import teamworkbranch.exception.NonprivilegedUserException;
+import teamworkbranch.exception.NotExistedException;
 import teamworkbranch.module.group.dao.GMemberMapper;
 import teamworkbranch.module.group.dao.GroupMapper;
 import teamworkbranch.module.group.model.GMember;
@@ -45,10 +47,19 @@ public class GroupServiceImpl implements GroupService {
     /**
      * 删除团队
      * @param groupId
+     * @param memberName
      * @return
      */
-    public boolean deleteGroup(int groupId) {
-        groupMapper.deleteGroup(groupId);
+    public boolean deleteGroup(int groupId,String memberName) throws NotExistedException, NonprivilegedUserException {
+        GMember gMember = gMemberMapper.selectById(groupId,memberName);
+        if(gMember==null){
+            throw new NotExistedException();
+        }else if(gMember.getIs_manager()==0){
+            throw new NonprivilegedUserException();
+        }
+        else {
+            groupMapper.deleteGroup(groupId);
+        }
         return true;
     }
 
@@ -59,8 +70,20 @@ public class GroupServiceImpl implements GroupService {
      * @param groupId
      * @return
      */
-    public boolean editGroup(String name,String info,int groupId) {
-        groupMapper.updateGroup(name,info,groupId);
+    public boolean editGroup(String name,String info,int groupId,String memberName) throws NotExistedException, NonprivilegedUserException {
+        Group group = groupMapper.selectById(groupId);
+        GMember gMember = gMemberMapper.selectById(groupId,memberName);
+        if(gMember==null){
+            throw new NotExistedException();
+        }else if(gMember.getIs_manager()==0){
+            throw new NonprivilegedUserException();
+        }
+        else {
+            group.setName(name);
+            group.setInfo(info);
+            groupMapper.updateGroup(group);
+
+        }
         return true;
     }
